+++ crankshaft-docker/CHANGELOG.md
@@ -8,6 +8,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Revised
+
+* Added retry logic with exponential backoff to Docker operations to handle
+  transient failures ([#69](https://github.com/stjude-rust-labs/crankshaft/pull/69)).
+
 ## 0.3.0 - 09-03-2025
 
 ### Added